Fixed some bugs
The nav bar links were not centered. And the width of the projects and skills were increased. Media query was set for the padding and margins of section>div. The dimension of the profile image on tablet and mobile wasn't correct . And other minute bugs. They currently have been fixed.